backend/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-CORS
Flask-Limiter
python-dotenv
bleach

# Testing dependencies
pytest
pytest-cov
pytest-flask
